pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sintetizador-decomp"
dynamic = ["version"]
dependencies = [
    "pandas",
    "polars>=1.0.0",
    "pyarrow>=18.0.0",
    "click>=8.1.7",
    "idecomp>=1.9.0",
]
requires-python = ">= 3.11"
authors = [
  {name = "Rogerio Alves", email = "rogerioalves.ee@gmail.com"},
]
description = "sintetizador-decomp"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "ruff",
    "mypy",
    "types-python-dateutil",
    "furo",
    "sphinx-gallery",
    "sphinx",
    "numpydoc",
    "plotly",
    "matplotlib",
]

[project.urls]
Documentation = "https://rjmalves.github.io/sintetizador-decomp/"
Repository = "https://github.com/rjmalves/sintetizador-decomp/"

[project.scripts]
sintetizador-decomp = "main:main"

[tool.hatch.version]
path = "app/__init__.py"

[tool.hatch.build.targets.wheel]
include = [
  "main.py",
  "app/",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["idecomp.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["cfinterface.*"]
ignore_errors = true

# idecomp v1.9+ .read() classmethods return base types (SectionFile,
# BlockFile, RegisterFile) instead of Self, causing many false-positive
# errors throughout this module.
[[tool.mypy.overrides]]
module = ["app.adapters.repository.files"]
disable_error_code = ["assignment", "return-value", "attr-defined", "unused-ignore"]

[[tool.mypy.overrides]]
module = ["dateutil.*"]
ignore_errors = true

